IBM Machine Learning/Upload_To_IBM_Buckets/requirements.txt
astroid==2.2.5
botocore==1.12.183
docutils==0.14
ibm-cos-sdk==2.5.1
ibm-cos-sdk-core==2.5.1
ibm-cos-sdk-s3transfer==2.5.1
isort==4.3.21
jmespath==0.9.4
lazy-object-proxy==1.4.1
mccabe==0.6.1
Pillow==6.1.0
pylint==2.3.1
python-dateutil==2.8.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
